put Radio and RadioCard in the same file
