+++ libs/ui/index.ts
@@ -14,7 +14,6 @@ export * from './lib/multi-select/MultiSelect'
 export * from './lib/PageHeader'
 export * from './lib/progress/Progress'
 export * from './lib/radio/Radio'
-export * from './lib/radio-card/RadioCard'
 export * from './lib/radio-group/RadioGroup'
 export * from './lib/tabs/Tabs'
 export * from './lib/table/Table'
