+++ libs/ui/lib/radio-group/__stories__/RadioGroup.stories.mdx
@@ -1,7 +1,6 @@
 import { ArgsTable, Canvas, Story, Meta } from '@storybook/addon-docs'
 import { Default, DefaultColumn, Cards, Disabled } from './RadioGroup.stories'
 import { RadioGroup } from '../RadioGroup'
-import { RadioCard } from '../../radio-card/RadioCard'
 
 <Meta title="Components/RadioGroup" component={RadioGroup} />
