+++ libs/ui/lib/radio/__stories__/RadioCard.stories.mdx
@@ -7,7 +7,7 @@ import {
   TwoLinesChecked,
   TwoLinesDefault,
 } from './RadioCard.stories'
-import { RadioCard } from '../RadioCard'
+import { RadioCard } from '../Radio'
 
 <Meta
   title="Components/RadioCard"
